Rename `db' package to `backend'.
